+++ lib/debug/observatory_debugger.dart
@@ -490,7 +490,6 @@ class ObservatoryIsolate extends DebugIsolate {
 
   String get name => isolateRef.name;
 
-
   // Isolate names are something like `foo_app.dart$main`. This cleans them up
   // slightly to `foo_app.dart:main()`.
   String get displayName => name.contains(r'$') ? "${name.replaceAll(r'$', ':')}()" : name;